+++ CHANGES.md
@@ -12,3 +12,4 @@
 - Added support for displaying certain types differently in the SQLResultsTable.
 - Added a top bar with a logo
 - Added builds for deploying as npm package
+- Added a status bar and calculating cluster data status